feat(routes): extract logs blueprint
